[#9] fix: permitAll url 수정
에러 원인: PatternParseException
org.springframework.web.util.pattern.PatternParseException: No more pattern data allowed after {*...} or ** pattern element
