+++ src/main/java/com/baedal/owner/adapter/in/web/controller/OwnerController.java
@@ -1,15 +1,8 @@
 package com.baedal.owner.adapter.in.web.controller;
 
 import com.baedal.owner.adapter.in.web.dto.request.AddStoreRequest;
-import com.baedal.owner.adapter.in.web.dto.request.LoginRequest;
-import com.baedal.owner.adapter.in.web.dto.request.SignUpRequest;
-import com.baedal.owner.adapter.in.web.dto.response.LoginResponse;
-import com.baedal.owner.adapter.in.web.dto.response.SignUpResponse;
 import com.baedal.owner.adapter.in.web.mapper.OwnerWebMapper;
 import com.baedal.owner.application.command.AddStoreCommand;
-import com.baedal.owner.application.command.LoginCommand;
-import com.baedal.owner.application.command.SignupCommand;
-import com.baedal.owner.application.port.in.OwnerAuthenticateUseCase;
 import com.baedal.owner.application.port.in.OwnerUseCase;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -30,22 +23,6 @@ public class OwnerController {
   private final OwnerWebMapper mapper;
 
   private final OwnerUseCase ownerUseCase;
-
-  private final OwnerAuthenticateUseCase authenticateUseCase;
-
-  @PostMapping("/login")
-  public ResponseEntity<LoginResponse> login(@RequestBody LoginRequest req) {
-    LoginCommand.Request command = mapper.loginToCommand(req);
-    LoginCommand.Response commandResponse = authenticateUseCase.login(command);
-    return ResponseEntity.ok(mapper.loginToResponse(commandResponse));
-  }
-
-  @PostMapping("/signup")
-  public ResponseEntity<SignUpResponse> signUp(@RequestBody SignUpRequest request) {
-    SignupCommand.Request command = mapper.signupToCommand(request);
-    SignupCommand.Response commandResponse = authenticateUseCase.signUp(command);
-    return ResponseEntity.ok(mapper.signupToResponse(commandResponse));
-  }
 
   @PreAuthorize("hasRole('OWNER')")
   @PostMapping("/addStore")
